Create config.js
My first addition of customization to config.js file (base taken from config.js.sample). Added my google calendar in the top right corner.

I can run the server using npm start, but it is not running properly. This has something to do with running this on Windows 10 (should be fine on Linux).